Module google_localservices1::api

Structs

AccountReportMethods
A builder providing access to all methods supported on accountReport resources. It is not used directly, but through the Localservices hub.

AccountReportSearchCall
Get account reports containing aggregate account data of all linked GLS accounts. Caller needs to provide their manager customer id and the associated auth credential that allows them read permissions on their linked accounts.

DetailedLeadReportMethods
A builder providing access to all methods supported on detailedLeadReport resources. It is not used directly, but through the Localservices hub.

DetailedLeadReportSearchCall
Get detailed lead reports containing leads that have been received by all linked GLS accounts. Caller needs to provide their manager customer id and the associated auth credential that allows them read permissions on their linked accounts.

GoogleAdsHomeservicesLocalservicesV1AccountReport
An Account Report of a GLS account identified by their account id containing aggregate data gathered from a particular date range. Next ID: 18

GoogleAdsHomeservicesLocalservicesV1AggregatorInfo
Conatiner for aggregator specific information if lead is for an aggregator GLS account.

GoogleAdsHomeservicesLocalservicesV1DetailedLeadReport
A Detailed Lead Report of a lead identified by their lead id and contains consumer, account, monetization, and lead data.

GoogleAdsHomeservicesLocalservicesV1MessageLead
Container for message lead specific information.

GoogleAdsHomeservicesLocalservicesV1PhoneLead
Container for phone lead specific information.

GoogleAdsHomeservicesLocalservicesV1SearchAccountReportsResponse
A page of the response received from the SearchAccountReports method. A paginated response where more pages are available has next_page_token set. This token can be used in a subsequent request to retrieve the next request page.

GoogleAdsHomeservicesLocalservicesV1SearchDetailedLeadReportsResponse
A page of the response received from the SearchDetailedLeadReports method. A paginated response where more pages are available has next_page_token set. This token can be used in a subsequent request to retrieve the next request page.

GoogleTypeTimeZone
Represents a time zone from the IANA Time Zone Database.

Localservices
Central instance to access all Localservices related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.